:recycle: refactor dir
