+++ aworld/workspace/artifact.py
